+++ CRUDTestProject/Data/Repositories/IBackupRepository.cs
@@ -1,0 +1,6 @@
+﻿namespace CRUDTestProject.Data.Repositories
+{
+    public interface IBackupRepository
+    {
+    }
+}
