+++ CRUDTestProject/Data/Repositories/Implementation/BackupRepository.cs
@@ -1,0 +1,6 @@
+﻿namespace CRUDTestProject.Data.Repositories.Implementation
+{
+    public class BackupRepository : IBackupRepository
+    {
+    }
+}
