+++ CRUDTestProject/Services/IBackupHandler.cs
@@ -1,0 +1,6 @@
+﻿namespace CRUDTestProject.Services
+{
+    public interface IBackupHandler
+    {
+    }
+}
