+++ MessagesTests/Services/BadWordsHandlerTests.cs
@@ -2,12 +2,8 @@
 using CRUDTestProject.Data.Repositories;
 using CRUDTestProject.Middleware.Exceptions;
 using CRUDTestProject.Services;
+using CRUDTestProject.Services.Implementation;
 using Moq;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace MessagesTests.Services
 {
