+++ MessagesTests/Services/MessageHandlerTests.cs
@@ -2,8 +2,7 @@
 using CRUDTestProject.Data.Repositories;
 using CRUDTestProject.Middleware.Exceptions;
 using CRUDTestProject.Services;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
+using CRUDTestProject.Services.Implementation;
 using Moq;
 
 namespace MessagesTests.Repository
